indexargs_url force ids to int
ids appear as hyperdb.String instances, which confused indexargs_url when they
appear in the filterspec. They need to be treated as treated as integers when
generating URLs.

It feels sort of hacky to check for 'id' like this but I'm at a loss for what
else to do in this case. Suggestions are welcome :)

Maybe we should look into using some other hyperdb class to represent ids?

this fixes 1783492

Some trailing whitespace also got trimmed.